refactored extension.ts into commands.ts contextStore.ts and added fixtures
