+++ src/fixtures/defaultPromptMarkdown.md
@@ -1,0 +1,7 @@
+---
+name:
+type: 
+---
+Fill in prompt text here.
+---
+Fill in response text here.
