+++ src/fixtures/defaultTreatment.treatments.yaml
@@ -1,0 +1,15 @@
+introSequences:
+  - name: "exampleIntro"
+    introSteps:
+      - name: "exampleStep1"
+        elements:
+          - type: "prompt"
+            
+treatments:
+  - name: "exampleTreatment"
+    playerCount: 1
+    gameStages:
+      - name: "exampleStage1"
+        duration: 60
+        elements:
+          - type: "prompt"
